Revert "Fix for the workaround for the Chinese SATA network adaptor: I forgot that"
This reverts commit 4b23cb257639f02e2152c6d116478a13770f717c.

I modified PS2Ident to show the entire 16-bit revision register, and it
reports its revision as 0x00FF. Rolling back to the previous revision
because that revision was working correctly.

It IS debatable whether it's worth keeping a workaround for this broken
adaptor, but I've decided that there's no harm in providing a workaround
for it. If someday, somebody releases a good SATA network adaptor, please
feel free to entirely remove this workaround!