Fixed program closing and deinitialization: 1. crt0 did not have everything that was related to loader support removed: it still involked the ExitDeleteThread() syscall, which prevented the program from quitting to the OSD. 2. libc will now no longer automatically close all files. While it seems like a good idea, it's not safe because it requires the SIF after the program has deinitialized it.
Although the SIF does get reinitialized when the I/O library communicates with the IOP again, it's not safe because the program has quit and the region that it exists at might get wiped/overwritten. If the EE kernel executes the old SIF handler from that region, the kernel might crash.