Disabled alignment of the payload field within pbuf_alloc for PBUF_RAM, fixed unaligned setting of the TCP packet options and changed output code of ps2ip-nm to only copy if necessary.
For the PS2, alignment of the payload field (for only PBUF_RAM) is undesirable because that would leave the final address of payload unaligned (aligned at some boundary that is a multiple of 2).
However, LWIP was also made to assume that the TCP header is aligned - which then becomes false if the payload field stops being aligned.