(EE LWIP) Changed sys_arch_protect to always invoke DIntr() and increased PBUF count to 80.
I am not sure if it will preserve the value of ProtLevel across the call to DIntr, but it should not. This will ensure that it will surely not.
As mentioned in the previous commit, the IOP NETMAN implementation is known to encounter cases of the EE running out of PBUF space, so the number of PBUFs have been further increased.