Corrected declaration (alignment) of SIFCMD structures within libkernel and changed SifInitCmd to use struct ca_pkt instead of an array.
Implicit alignment of 16 bytes is not done in the SCE libs and may be for a good reason; 16-byte alignment can be insufficient, depending on the design of the code.
When data is written back on an unaligned address and is mixed with the use of the uncached segment, the data in RAM can become corrupted by the out-of-date values in the cache.

Lastly, declaring the IOP reset packet structure as aligned caused its size to be also aligned, which then deviates from its original definition (0x70 bytes instead of 0x68).