(HDPro) added missing check on nsector and code from the late ATAD v2.4 driver and improved on comments/code cosmetics.
The check on nsector is missing in the original HDPro driver, but the ATA-4 specification (9.1) did indicate that sector should be 1.
So it should be safe to check, as in the normal ATAD module.

As for the check on lcyl = 55h, hcyl = AAh, wisi told me that it might be to check on whether the ATA device is connected properly.