(PFS) Changed UID and GID to 0xFFFF, hardcoded. Removed option to change UID and GID via chstat. Removed unofficial option to set UID and GID.
As of r2.3, Sony already changed the UID and GID to 0xFFFF.
The HDD browser might have been built with an older library, which has UID and GID to set 0, as documented.
As folders created by the HDD Browser have UID & GID set to 0xFFFF, it might be doing so through chstat().

Software may ignore the entry if the UID and GID are not 0xFFFF.