Move system clock to TIM2
TIM2 is not used and, in theory, reserved for the kernel (but it never
uses it). Since we only use it to measure the time, it's easy to
configure it (using kmode enter/exit to get around permissions) and use
it without much overhead, while at the same time protecting it from the
users.

This will free timer 0 and 1 to be used by homebrew users, and
potentially by the timer library in PR #266