Remove bin2s and bin2o in favor of bin2c
Creating .s and .o files is much more complicated than creating a .c
file. So instead of trying to maintain these 2 complex variants (that
now seem to have issues), we leave the complex part of generating .s and
.o files upto the toolchain.
We have bin2c, and with the .c file you can let the toolchain generate
.s and .o files if needed. Always up-to-date with whatever toolchain you
use.